Extract SpecifiedRuleViolation into separate class
